Revert "ARROW-7863: [C++][Python][CI] Ensure running HDFS related tests" [DNM] fix rebase
This reverts commit 4e53749097ba687afd5e000067925def2e2802c9.

+++ java/adapter/common/src/main/java/org/apache/arrow/adapter/common/AdaptorReferenceManager.java
@@ -92,7 +92,7 @@ public class AdaptorReferenceManager implements ReferenceManager {
   }
 
   @Override
-  public ArrowBuf deriveBuffer(ArrowBuf sourceBuffer, int index, int length) {
+  public ArrowBuf deriveBuffer(ArrowBuf sourceBuffer, long index, long length) {
     final long derivedBufferAddress = sourceBuffer.memoryAddress() + index;
 
     // create new ArrowBuf
@@ -117,12 +117,12 @@ public class AdaptorReferenceManager implements ReferenceManager {
   }
 
   @Override
-  public int getSize() {
+  public long getSize() {
     return size;
   }
 
   @Override
-  public int getAccountedSize() {
+  public long getAccountedSize() {
     return 0;
   }
 }
